Chore: decouple migrate lockfile (#6107)
