+++ vscode/microsoft-kiota/src/commands/migrate/migrateFromLockFileCommand.ts
@@ -2,8 +2,8 @@ import * as vscode from "vscode";
 import { ExtensionContext, Uri, workspace } from "vscode";
 
 import { extensionId } from "../../constants";
-import { handleMigration } from "../../util";
 import { Command } from "../Command";
+import { handleMigration } from "./migrateFromLockFile.util";
 
 export class MigrateFromLockFileCommand extends Command {
   private _context: ExtensionContext;
