+++ vscode/microsoft-kiota/src/extension.ts
@@ -8,7 +8,7 @@ import { DeleteWorkspaceItemCommand } from './commands/deleteWorkspaceItem/delet
 import { EditPathsCommand } from './commands/editPathsCommand';
 import { GenerateClientCommand } from './commands/generate/generateClientCommand';
 import { displayGenerationResults } from './commands/generate/generation-util';
-import { checkForLockFileAndPrompt } from "./commands/migrate/migrateFromLockFile";
+import { checkForLockFileAndPrompt } from "./commands/migrate/migrateFromLockFile.util";
 import { MigrateFromLockFileCommand } from './commands/migrate/migrateFromLockFileCommand';
 import { SearchOrOpenApiDescriptionCommand } from './commands/openApidescription/searchOrOpenApiDescriptionCommand';
 import { AddAllToSelectedEndpointsCommand } from './commands/openApiTreeView/addAllToSelectedEndpointsCommand';
