+++ vscode/microsoft-kiota/src/util.ts
@@ -3,7 +3,6 @@ import * as path from 'path';
 import * as vscode from 'vscode';
 import { QuickPickItem } from "vscode";
 
-import { displayMigrationMessages, migrateFromLockFile } from './commands/migrate/migrateFromLockFile';
 import { APIMANIFEST, CLIENT, CLIENTS, KIOTA_DIRECTORY, KIOTA_WORKSPACE_FILE, PLUGIN, PLUGINS } from './constants';
 import { allGenerationLanguages } from './kiotaInterop';
 import { GenerationType, KiotaGenerationLanguage, KiotaPluginType } from './types/enums';
@@ -69,33 +68,6 @@ export function findAppPackageDirectory(directory: string): string | null {
   }
 
   return null;
-}
-
-export async function handleMigration(
-  context: vscode.ExtensionContext,
-  workspaceFolder: vscode.WorkspaceFolder
-): Promise<void> {
-  vscode.window.withProgress({
-      location: vscode.ProgressLocation.Notification,
-      title: vscode.l10n.t("Migrating your API clients..."),
-      cancellable: false
-  }, async (progress) => {
-      progress.report({ increment: 0 });
-
-      try {
-          const migrationResult = await migrateFromLockFile(context, workspaceFolder.uri.fsPath);
-
-          progress.report({ increment: 100 });
-
-          if (migrationResult && migrationResult.length > 0) {
-              displayMigrationMessages(migrationResult);
-          } else {
-              vscode.window.showWarningMessage(vscode.l10n.t("Migration completed, but no changes were detected."));
-          }
-      } catch (error) {
-          vscode.window.showErrorMessage(vscode.l10n.t(`Migration failed: ${error}`));
-      }
-  });
 }
 
 export function getSanitizedString(rawValue?: string): string| undefined{
